Simple pattern comparison program
This program was programmed using Erlang , i have created 3 functions in
this program .One function "xOr(A,B)" checks the user input and compares
A to B and gives a result according to its findings. Second function
"maxThree(A,B,C)"  takes three peices of data from the user and checks
which of the 3 values is the biggest .The third function
"howManyEqual(X,Y,Z)" takes in 3 user inputs and checks how many of the
3 values are equal and dispalys a number if all values are different for
example a 0 will be displayed.